backend/requirements.txt
# Django Backend Requirements
# Logis Career AI Platform

# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Static files
whitenoise>=6.6.0

# Image handling
Pillow>=10.0.0

# Development
python-dotenv>=1.0.0

# Production
gunicorn>=21.2.0
